+++ src/tools/quickcode.ts
@@ -572,6 +572,7 @@ ${ $h.toStringSafe(DOCUMENT.getText()) }
                             break;
 
                         case 'html':
+                        case 'xml':
                             {
                                 const beautify = require('js-beautify').html;
 
@@ -628,6 +629,19 @@ ${ $h.toStringSafe(DOCUMENT.getText()) }
                                 result = beautify(result);
 
                                 lang = 'javascript';
+                            }
+                            break;
+
+                        case 'jade':
+                            {
+                                const beautify = require('js-beautify').html;
+                                const pug = require('pug');
+
+                                result = beautify(
+                                    pug.render(TEXT)
+                                );
+
+                                lang = 'html';
                             }
                             break;
 
@@ -697,6 +711,23 @@ ${ $h.toStringSafe(DOCUMENT.getText()) }
                                 result = new cleanCSS()
                                     .minify(TEXT)
                                     .styles;
+                            }
+                            break;
+
+                        case 'html':
+                        case 'xml':
+                            {
+                                const htmlMinify = require('html-minifier');
+
+                                result = htmlMinify.minify(TEXT, {
+                                    html5: true,
+                                    minifyCSS: true,
+                                    minifyJS: true,
+                                    removeComments: true,
+                                    preserveLineBreaks: false,
+                                    collapseWhitespace: true,
+                                    keepClosingSlash: true,
+                                });
                             }
                             break;
 
@@ -793,6 +824,15 @@ ${ $h.toStringSafe(DOCUMENT.getText()) }
         }
     );
 
+    // @ts-ignore
+    const $base64 = asAsync_628dffd9c1e74e5cb82620a2c575e5dd(
+        async (val: any, enc?: string) => {
+            return (
+                await $buff(val, enc)
+            ).toString('base64');
+        }
+    );
+
     // code to execute
     let _code_g93c97d35bd94b22b3041037bdc64780: string = $h.toStringSafe(_opts_f4eba53df3b74b7aa4e3a3228b528d78.code);
     if (!_code_g93c97d35bd94b22b3041037bdc64780.trim().startsWith('return ')) {
@@ -885,6 +925,7 @@ async function showHelp_579c52a1992b472183db2fff8c764504() {
         md += '---- | ----------- | -------\n';
         md += '`$alert(msg)` | Shows a (warning) popup. | `$alert("Hello, TM!")`\n';
         md += '`$asc(str)` | Handles a value as string and returns the ASCII (codes). | `$asc("T")`\n';
+        md += '`$base64(val, enc?)` | Converts a value to a Base64 string. | `$base64("mkloubert:P@ssword123!", "ascii")`\n';
         md += '`$beautify` | Beautifies the code in the active editor and opens the result in a new one. | `$beautify`\n';
         md += '`$buff(val, enc?)` | Converts a value to a [Buffer](https://nodejs.org/api/buffer.html), if needed. | `$buff("TM")`\n';
         md += '`$cmd(id, ...args)` | Executes a [Visual Studio Code command](https://code.visualstudio.com/api/references/commands). | `$cmd("vscode.openFolder")`\n';